Don't try to skip over text we've parsed
Since mColorCursor is 0 in the dormant state, the EditorText was incorrectly
queueing edits in the dormant state since every index is ahead of 0.

This commit fixes that by adding an extra check to replace() to make sure
we're not dormant before we go about queueing edits. It also introduces -1 instead
of 0 as the dormant state sentinel value, since it's feasible the user could
edit text after the file was parsed but before mColorCursor gets to be touched.